refactor: extract runtime output utilities into tau-runtime
