+++ crates/tau-coding-agent/src/runtime_output.rs
@@ -1,36 +1,14 @@
 use super::*;
 
 pub(crate) fn summarize_message(message: &Message) -> String {
-    let text = message.text_content().replace('\n', " ");
-    if text.trim().is_empty() {
-        return format!(
-            "{:?} (tool_calls={})",
-            message.role,
-            message.tool_calls().len()
-        );
-    }
-
-    let max = 60;
-    if text.chars().count() <= max {
-        text
-    } else {
-        let summary = text.chars().take(max).collect::<String>();
-        format!("{summary}...")
-    }
+    tau_runtime::runtime_output_runtime::summarize_message(message)
 }
 
 pub(crate) fn persist_messages(
     session_runtime: &mut Option<SessionRuntime>,
     new_messages: &[Message],
 ) -> Result<()> {
-    let Some(runtime) = session_runtime.as_mut() else {
-        return Ok(());
-    };
-
-    runtime.active_head = runtime
-        .store
-        .append_messages(runtime.active_head, new_messages)?;
-    Ok(())
+    tau_runtime::runtime_output_runtime::persist_messages(session_runtime, new_messages)
 }
 
 pub(crate) fn print_assistant_messages(
@@ -38,97 +16,19 @@ pub(crate) fn print_assistant_messages(
     render_options: RenderOptions,
     suppress_first_streamed_text: bool,
 ) {
-    let mut suppressed_once = false;
-    for message in messages {
-        if message.role != MessageRole::Assistant {
-            continue;
-        }
-
-        let text = message.text_content();
-        if !text.trim().is_empty() {
-            if render_options.stream_output && suppress_first_streamed_text && !suppressed_once {
-                suppressed_once = true;
-                println!("\n");
-                continue;
-            }
-            println!();
-            if render_options.stream_output {
-                let mut stdout = std::io::stdout();
-                for chunk in stream_text_chunks(&text) {
-                    print!("{chunk}");
-                    let _ = stdout.flush();
-                    if render_options.stream_delay_ms > 0 {
-                        std::thread::sleep(Duration::from_millis(render_options.stream_delay_ms));
-                    }
-                }
-                println!("\n");
-            } else {
-                println!("{text}\n");
-            }
-            continue;
-        }
-
-        let tool_calls = message.tool_calls();
-        if !tool_calls.is_empty() {
-            println!(
-                "\n[assistant requested {} tool call(s)]\n",
-                tool_calls.len()
-            );
-        }
-    }
+    tau_runtime::runtime_output_runtime::print_assistant_messages(
+        messages,
+        render_options.stream_output,
+        render_options.stream_delay_ms,
+        suppress_first_streamed_text,
+    );
 }
 
+#[cfg(test)]
 pub(crate) fn stream_text_chunks(text: &str) -> Vec<&str> {
-    text.split_inclusive(char::is_whitespace).collect()
+    tau_runtime::runtime_output_runtime::stream_text_chunks(text)
 }
 
 pub(crate) fn event_to_json(event: &AgentEvent) -> serde_json::Value {
-    match event {
-        AgentEvent::AgentStart => serde_json::json!({ "type": "agent_start" }),
-        AgentEvent::AgentEnd { new_messages } => {
-            serde_json::json!({ "type": "agent_end", "new_messages": new_messages })
-        }
-        AgentEvent::TurnStart { turn } => serde_json::json!({ "type": "turn_start", "turn": turn }),
-        AgentEvent::TurnEnd {
-            turn,
-            tool_results,
-            request_duration_ms,
-            usage,
-            finish_reason,
-        } => serde_json::json!({
-            "type": "turn_end",
-            "turn": turn,
-            "tool_results": tool_results,
-            "request_duration_ms": request_duration_ms,
-            "usage": usage,
-            "finish_reason": finish_reason,
-        }),
-        AgentEvent::MessageAdded { message } => serde_json::json!({
-            "type": "message_added",
-            "role": format!("{:?}", message.role).to_lowercase(),
-            "text": message.text_content(),
-            "tool_calls": message.tool_calls().len(),
-        }),
-        AgentEvent::ToolExecutionStart {
-            tool_call_id,
-            tool_name,
-            arguments,
-        } => serde_json::json!({
-            "type": "tool_execution_start",
-            "tool_call_id": tool_call_id,
-            "tool_name": tool_name,
-            "arguments": arguments,
-        }),
-        AgentEvent::ToolExecutionEnd {
-            tool_call_id,
-            tool_name,
-            result,
-        } => serde_json::json!({
-            "type": "tool_execution_end",
-            "tool_call_id": tool_call_id,
-            "tool_name": tool_name,
-            "is_error": result.is_error,
-            "content": result.content,
-        }),
-    }
+    tau_runtime::runtime_output_runtime::event_to_json(event)
 }
