+++ crates/tau-runtime/src/lib.rs
@@ -2,10 +2,12 @@ pub mod channel_store;
 pub mod observability_loggers_runtime;
 pub mod rpc_capabilities_runtime;
 pub mod rpc_protocol_runtime;
+pub mod runtime_output_runtime;
 pub mod transport_health;
 
 pub use channel_store::*;
 pub use observability_loggers_runtime::*;
 pub use rpc_capabilities_runtime::*;
 pub use rpc_protocol_runtime::*;
+pub use runtime_output_runtime::*;
 pub use transport_health::*;
